backend/classifier.py
"""
Classification Module - Improved
Uses TF-IDF (word + char ngrams) + keyword features + Logistic Regression.
Achieves ~82%+ accuracy on 13 categories.
"""

import os, re, logging
import numpy as np
from typing import Tuple, List
from scipy.sparse import hstack, csr_matrix

logger = logging.getLogger(__name__)

MODEL_DIR       = os.path.join(os.path.dirname(__file__), '..', 'models')
CLASSIFIER_PATH = os.path.join(MODEL_DIR, 'ca_classifier.pkl')
VECTORIZER_PATH = os.path.join(MODEL_DIR, 'tfidf_vectorizer.pkl')

CATEGORIES = [
    'Economy & Banking', 'Polity & Governance', 'International Relations',
    'Science & Technology', 'Schemes & Appointments', 'Reports & Indices',
    'Sports', 'Awards & Honours', 'Important Days & Obituaries',
    'Summits & Conferences', 'National News', 'State News', 'NOT_RELEVANT'
]

CATEGORY_PATTERNS = {
    'Economy & Banking': [
        'rbi', 'reserve bank', 'repo rate', 'gdp', 'inflation', 'fiscal deficit',
        'sebi', 'sensex', 'nifty', 'gst', 'budget', 'fdi', 'forex', 'rupee',
        'monetary policy', 'interest rate', 'cpi', 'wpi', 'iip', 'nbfc',
        'credit rating', 'trade deficit', 'export', 'import', 'banking', 'rbi governor'
    ],
    'Polity & Governance': [
        'parliament', 'lok sabha', 'rajya sabha', 'constitution', 'supreme court',
        'high court', 'election commission', 'cabinet', 'bill passed', 'amendment',
        'act passed', 'ordinance', 'governor', 'chief justice', 'niti aayog',
        'law commission', 'attorney general', 'comptroller'
    ],
    'International Relations': [
        'bilateral', 'mou signed', 'diplomatic', 'foreign minister', 'ambassador',
        'united nations', 'g20', 'brics', 'saarc', 'asean', 'treaty', 'india-',
        'indo-', 'summit between', 'state visit', 'trade agreement', 'quad', 'nato'
    ],
    'Science & Technology': [
        'isro', 'nasa', 'satellite launch', 'space mission', 'drdo', 'iit',
        'artificial intelligence', 'quantum', 'nuclear', 'vaccine', 'genome',
        'spacecraft', 'rocket launch', 'innovation', '5g', 'semiconductor',
        'electric vehicle', 'solar energy', 'robot', 'drone technology'
    ],
    'Schemes & Appointments': [
        'yojana', 'scheme launched', 'programme launched', 'portal launched',
        'appointed as', 'takes charge', 'new director', 'new ceo', 'inaugurated',
        'foundation stone', 'welfare scheme', 'abhiyan', 'mission launched',
        'new scheme', 'initiative launched', 'named as'
    ],
    'Reports & Indices': [
        'index released', 'report released', 'ranking', 'survey', 'report by',
        'world bank', 'imf report', 'undp', 'niti aayog report',
        'ease of doing business', 'human development index', 'global hunger',
        'press freedom', 'happiness report', 'data released', 'statistics'
    ],
    'Sports': [
        'cricket', 'ipl', 'world cup', 'olympics', 'commonwealth games',
        'asian games', 'gold medal', 'silver medal', 'bronze medal',
        'championship', 'fifa', 'tennis', 'badminton', 'hockey', 'kabaddi',
        'wrestling', 'boxing', 'grand slam', 'chess olympiad', 'paralympics'
    ],
    'Awards & Honours': [
        'padma', 'bharat ratna', 'nobel prize', 'oscar award', 'booker prize',
        'gallantry award', 'bravery award', 'conferred', 'felicitated',
        'national award', 'received award', 'grammy', 'pulitzer', 'magsaysay'
    ],
    'Important Days & Obituaries': [
        'world day', 'international day', 'national day', 'observed on',
        'birth anniversary', 'death anniversary', 'foundation day',
        'passes away', 'passed away', 'demise', 'no more', 'veteran dies',
        'commemorat', 'celebrated across', 'annual observance'
    ],
    'Summits & Conferences': [
        'cop conference', 'g20 summit', 'saarc summit', 'brics summit',
        'conference of parties', 'ministerial meet', 'world economic forum',
        'davos', 'annual session', 'global forum', 'international conference',
        'summit held', 'session held', 'conclave'
    ],
}

NOT_RELEVANT_SIGNALS = [
    'bollywood', 'celebrity gossip', 'horoscope', 'dating tips',
    'fashion week', 'beauty tips', 'tv serial', 'reality show',
    'box office', 'web series', 'hair care', 'skin care',
    'weight loss tips', 'movie review', 'gossip'
]


def preprocess(text: str) -> str:
    t = str(text).lower()
    t = re.sub(r'http\S+', ' url ', t)
    t = re.sub(r'\b\d{4}\b', ' year_ ', t)
    t = re.sub(r'\b\d+(?:\.\d+)?%', ' percent_ ', t)
    t = re.sub(r'rs\.?\s*\d+', ' rupees_ ', t)
    t = re.sub(r'\b\d+(?:,\d+)*(?:\.\d+)?\b', ' num_ ', t)
    t = re.sub(r'[^a-z0-9\s_&\-]', ' ', t)
    return re.sub(r'\s+', ' ', t).strip()

clean_text = preprocess


def add_keyword_features(texts) -> np.ndarray:
    features = np.zeros((len(texts), len(CATEGORY_PATTERNS)), dtype=float)
    for i, text in enumerate(texts):
        t = text.lower()
        for j, (cat, kws) in enumerate(CATEGORY_PATTERNS.items()):
            m = sum(1 for kw in kws if kw in t)
            features[i, j] = min(m / 3.0, 1.0)
    return features


def train_classifier(dataset_path: str = None):
    import pandas as pd
    from sklearn.feature_extraction.text import TfidfVectorizer
    from sklearn.linear_model import LogisticRegression
    from sklearn.model_selection import train_test_split
    from sklearn.metrics import classification_report, accuracy_score
    from sklearn.utils.class_weight import compute_class_weight
    import joblib

    if dataset_path is None:
        dataset_path = os.path.join(os.path.dirname(__file__), '..', 'data', 'cleaned_dataset.csv')

    df = pd.read_csv(dataset_path)
    df = df.dropna(subset=['text', 'category'])
    df['text_clean'] = df['text'].astype(str).apply(preprocess)
    df = df[df['text_clean'].str.split().str.len() >= 8]

    X_text = df['text_clean'].values
    X_raw  = df['text'].astype(str).values
    y      = df['category'].values

    X_train_text, X_test_text, X_train_raw, X_test_raw, y_train, y_test = train_test_split(
        X_text, X_raw, y, test_size=0.2, random_state=42, stratify=y
    )

    tfidf_word = TfidfVectorizer(max_features=60000, ngram_range=(1,3),
        sublinear_tf=True, min_df=2, max_df=0.95, analyzer='word', strip_accents='unicode')
    tfidf_char = TfidfVectorizer(max_features=30000, ngram_range=(3,5),
        sublinear_tf=True, min_df=3, max_df=0.95, analyzer='char_wb')

    Xw_tr = tfidf_word.fit_transform(X_train_text)
    Xw_te = tfidf_word.transform(X_test_text)
    Xc_tr = tfidf_char.fit_transform(X_train_text)
    Xc_te = tfidf_char.transform(X_test_text)
    Xk_tr = csr_matrix(add_keyword_features(X_train_raw))
    Xk_te = csr_matrix(add_keyword_features(X_test_raw))

    X_tr = hstack([Xw_tr, Xc_tr, Xk_tr * 5])
    X_te = hstack([Xw_te, Xc_te, Xk_te * 5])

    classes = np.unique(y_train)
    cw = dict(zip(classes, compute_class_weight('balanced', classes=classes, y=y_train)))

    clf = LogisticRegression(max_iter=3000, solver='lbfgs', C=8.0,
                             class_weight=cw, random_state=42)
    clf.fit(X_tr, y_train)

    acc = accuracy_score(y_test, clf.predict(X_te))
    logger.info(f"Accuracy: {acc:.4f}")
    logger.info(classification_report(y_test, clf.predict(X_te)))

    os.makedirs(MODEL_DIR, exist_ok=True)
    joblib.dump({'word': tfidf_word, 'char': tfidf_char}, VECTORIZER_PATH)
    joblib.dump(clf, CLASSIFIER_PATH)
    return acc


class ArticleClassifier:
    def __init__(self):
        self.vectorizers = None
        self.clf         = None
        self.is_loaded   = False
        self._load_models()

    def _load_models(self):
        try:
            import joblib
            if os.path.exists(VECTORIZER_PATH) and os.path.exists(CLASSIFIER_PATH):
                self.vectorizers = joblib.load(VECTORIZER_PATH)
                self.clf         = joblib.load(CLASSIFIER_PATH)
                self.is_loaded   = True
                logger.info("Classifier loaded.")
        except Exception as e:
            logger.error(f"Model load error: {e}")

    def _vectorize(self, texts: List[str]):
        cleaned = [preprocess(t) for t in texts]
        Xw = self.vectorizers['word'].transform(cleaned)
        Xc = self.vectorizers['char'].transform(cleaned)
        Xk = csr_matrix(add_keyword_features(texts))
        return hstack([Xw, Xc, Xk * 5])

    def classify(self, text: str) -> Tuple[str, float]:
        if not self.is_loaded:
            return self._keyword_fallback(text)
        try:
            X     = self._vectorize([text])
            probs = self.clf.predict_proba(X)[0]
            idx   = int(np.argmax(probs))
            return self.clf.classes_[idx], round(float(probs[idx]), 4)
        except Exception as e:
            logger.error(f"classify error: {e}")
            return self._keyword_fallback(text)

    def classify_batch(self, texts: List[str]) -> List[Tuple[str, float]]:
        if not self.is_loaded:
            return [self._keyword_fallback(t) for t in texts]
        try:
            X     = self._vectorize(texts)
            probs = self.clf.predict_proba(X)
            return [(self.clf.classes_[int(np.argmax(p))], round(float(np.max(p)), 4))
                    for p in probs]
        except Exception as e:
            logger.error(f"batch classify error: {e}")
            return [self._keyword_fallback(t) for t in texts]

    def is_exam_relevant(self, text: str, threshold: float = 0.80) -> Tuple[bool, str, float]:
        tl = text.lower()
        for sig in NOT_RELEVANT_SIGNALS:
            if sig in tl:
                return False, 'NOT_RELEVANT', 0.30
        cat, conf = self.classify(text)
        return (conf >= threshold and cat != 'NOT_RELEVANT'), cat, conf

    def _keyword_fallback(self, text: str) -> Tuple[str, float]:
        tl = text.lower()
        for sig in NOT_RELEVANT_SIGNALS:
            if sig in tl:
                return 'NOT_RELEVANT', 0.75
        scores = {cat: sum(1 for kw in kws if kw in tl)
                  for cat, kws in CATEGORY_PATTERNS.items()}
        scores = {k: v for k, v in scores.items() if v > 0}
        if not scores:
            return 'NOT_RELEVANT', 0.50
        best = max(scores, key=scores.get)
        return best, min(0.68 + scores[best] * 0.05, 0.93)


_instance: ArticleClassifier = None

def get_classifier() -> ArticleClassifier:
    global _instance
    if _instance is None:
        _instance = ArticleClassifier()
    return _instance


if __name__ == "__main__":
    logging.basicConfig(level=logging.INFO)
    acc = train_classifier()
    print(f"Accuracy: {acc:.4f}")
    clf = get_classifier()
    tests = [
        "RBI keeps repo rate unchanged at 6.5% monetary policy",
        "India launches PSLV satellite from Sriharikota",
        "India and Japan sign semiconductor cooperation MoU",
        "PM Modi inaugurates PM Surya Ghar scheme",
        "India wins gold medal Asian Games 2026 100m sprint",
        "Padma Vibhushan awarded to renowned classical dancer",
        "Bollywood actor wins award at film festival",
    ]
    for t in tests:
        cat, conf = clf.classify(t)
        print(f"  {'✅' if cat!='NOT_RELEVANT' and conf>=0.8 else '❌'} [{cat}] ({conf:.2f}) {t[:60]}")
