requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
apscheduler>=3.10.4
feedparser>=6.0.10
httpx>=0.25.0
requests>=2.31.0
trafilatura>=1.6.0
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0
nltk>=3.8.1
networkx>=3.1
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
aiofiles>=23.2.1
pydantic>=2.4.0